chore: remove some warinigs
